requirements.txt
pypdf==3.15.5
langchain==0.0.284
langsmith==0.0.33
openai==0.28.0
openpyxl==3.1.2
pandas==2.0.3
numpy==1.24.4
streamlit==1.26.0
streamlit-chat==0.1.1
python-dotenv==1.0.0
tiktoken==0.4.0
chromadb==0.4.9
openai==0.28.0
boto3==1.28.68
botocore==1.31.68
aiosqlite @ file:///croot/aiosqlite_1683773899903/work
sqlparse==0.4.4